Architecture/dataset unification

+++ utils.py
@@ -120,7 +120,7 @@ def normalize_features(mx):
 
 def accuracy(output, labels):
     preds = output.max(1)[1].type_as(labels)
-    correct = preds.eq(labels).double()
+    correct = preds.eq(labels)
     correct = correct.sum()
     return correct / len(labels)
